Extract shared HTTP response logic into BaseJSONRequestHandler
Moves `_json_response`, `_error_response`, and `request_id` from
GarminAccountLinkHandler and GoogleHealthAccountLinkHandler into a common
base class to eliminate code duplication.
